# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A secure, decentralized file sharing platform with end-to-end encryption, digital signatures, temporary links, and blockchain audit trails.",
        "title": "SafeTransfer API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/download/{cid}": {
            "get": {
                "description": "Downloads a file from IPFS using its CID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "summary": "Download a file",
                "operationId": "download-file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CID of the file to download",
                        "name": "cid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File data",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "Uploads a file to IPFS and returns the CID.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Uploads a file to IPFS",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/upload.FileUploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "upload.FileUploadResponse": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "string"
                }
            }
        }
    }
}